Explicitly cast stats to double before divide.
#rb none

#ROBOMERGE-AUTHOR: robert.millar
#ROBOMERGE-SOURCE: CL 17878830 via CL 17878891 via CL 17878901 via CL 17878909
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v883-17842818)

[CL 17878921 by robert millar in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/Core/Private/ProfilingDebugging/TraceAuxiliary.cpp
@@ -377,9 +377,9 @@ void FTraceAuxiliaryImpl::UpdateCsvStats() const
 		UE::Trace::FStatistics Stats;
 		UE::Trace::GetStatistics(Stats);
 
-		CSV_CUSTOM_STAT(Trace, MemoryUsedMb,	Stats.MemoryUsed / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
-		CSV_CUSTOM_STAT(Trace, CacheUsedMb,		Stats.CacheUsed / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
-		CSV_CUSTOM_STAT(Trace, CacheWasteMb,	Stats.CacheWaste / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
+		CSV_CUSTOM_STAT(Trace, MemoryUsedMb,	double(Stats.MemoryUsed) / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
+		CSV_CUSTOM_STAT(Trace, CacheUsedMb,		double(Stats.CacheUsed) / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
+		CSV_CUSTOM_STAT(Trace, CacheWasteMb,	double(Stats.CacheWaste) / 1024.0 / 1024.0, ECsvCustomStatOp::Set);
 	}
 #endif
 }
